delete button done, still need to figure out comment submission issue

+++ frontend/src/components/Comments/SingleComment.js
@@ -4,20 +4,18 @@ import { useDispatch, useSelector } from "react-redux";
 import { removeComment } from "../../store/comments";
 
 const SingleComment = ({ comment }) => {
-    const dispatch = useDispatch()
-    const currentUser = useSelector((state) => state.session.user);
-
-
-
-
+  const dispatch = useDispatch();
+  const currentUser = useSelector((state) => state.session.user);
 
   const handleDelete = (e) => {
     e.preventDefault();
-    console.log(comment.id, "TARGET VALUE");
 
     if (!currentUser) return window.alert("Please Login");
-
-    dispatch(removeComment(comment.id));
+    if (currentUser.id !== comment.userId) {
+      return window.alert("You do not have access to delete this comment");
+    } else {
+      dispatch(removeComment(comment.id));
+    }
   };
 
   return <button onClick={handleDelete}>Delete</button>;
